requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tqdm>=4.65.0
wandb>=0.15.0
Pillow>=9.5.0  # For image processing
numpy>=1.24.0
pandas>=1.5.0  # For reading dataset annotations
requests>=2.31.0  # For downloading dataset
datasets>=2.12.0  # For loading Flickr30k dataset 